Move ProjectRootCanonicalizer from Domain to Infrastructure
The class is a P/Invoke-backed path canonicalizer whose only caller
is BridgeTransportEndpoint in Infrastructure. Domain should hold no
platform interop, and no inner layer consumes it, so it moves next
to its caller as internal instead of gaining a single-use port
interface. The onion guard test now pins it to Infrastructure.
