+++ Assets/Tests/Editor/BridgeTransportEndpointTests.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using NUnit.Framework;
 
-using io.github.hatayama.UnityCliLoop.Domain;
+using io.github.hatayama.UnityCliLoop.Infrastructure;
 
 namespace io.github.hatayama.UnityCliLoop.Tests.Editor
 {
