+++ Packages/src/Editor/Infrastructure/ProjectRootCanonicalizer.cs.meta
